Flatten the config dependency tree
Big config header files should only depend on other (dumb) config
definitions and general types.

This way, we make sure that interpendencies between modules are
explicit and not transitively inherited from an include to `config/config.h`
