+++ src/libaktualizr/package_manager/ostreemanager.h
@@ -7,9 +7,9 @@
 #include <glib/gi18n.h>
 #include <ostree-1/ostree.h>
 
-#include "config/config.h"
 #include "crypto/keymanager.h"
-#include "package_manager/packagemanagerinterface.h"
+#include "packagemanagerconfig.h"
+#include "packagemanagerinterface.h"
 #include "utilities/types.h"
 
 const char remote[] = "aktualizr-remote";
